relocate burger in scroll-bar
