+++ commercialNetwork/src/js/scrollBar.js
@@ -14,104 +14,104 @@ let timeoutId;
 let clientY1;
 
 const watchedSlide = new Proxy(
-  {
-    activeSlide: 0,
-  },
-  onSlideChange()
+	{
+		activeSlide: 0,
+	},
+	onSlideChange()
 );
 
 function onSlideChange() {
-  return {
-    set(target, property, slideIndex) {
-      console.log(property);
-      console.log(target);
-      console.log(slideIndex);
-      if (slideIndex < 0 || slideIndex > 11) {
-        console.log(true);
-        return;
-      }
+	return {
+		set(target, property, slideIndex) {
+			console.log(property);
+			console.log(target);
+			console.log(slideIndex);
+			if (slideIndex < 0 || slideIndex > 11) {
+				console.log(true);
+				return;
+			}
 
-      const sectionHeight = document.documentElement.clientHeight;
+			const sectionHeight = document.documentElement.clientHeight;
 
-      counter.innerText = slideIndex + 1;
+			counter.innerText = slideIndex + 1;
 
-      main.style.transform = `translateY(${-(sectionHeight * slideIndex)}px)`;
+			main.style.transform = `translateY(${-(sectionHeight * slideIndex)}px)`;
 
-      sections.forEach((section) => section.classList.remove("active"));
-      indicatorLines.forEach((line) => line.classList.remove("active"));
-      indicator.classList.remove("black");
-      header.classList.remove("black");
+			sections.forEach((section) => section.classList.remove("active"));
+			indicatorLines.forEach((line) => line.classList.remove("active"));
+			indicator.classList.remove("black");
+			header.classList.remove("black");
 
-      sections[slideIndex].classList.add("active");
-      indicatorLines[slideIndex].classList.add("active");
+			sections[slideIndex].classList.add("active");
+			indicatorLines[slideIndex].classList.add("active");
 
-      if (BLACK_SECTION_INDEXES.includes(slideIndex)) {
-        indicator.classList.add("black");
-        header.classList.add("black");
-      }
+			if (BLACK_SECTION_INDEXES.includes(slideIndex)) {
+				indicator.classList.add("black");
+				header.classList.add("black");
+			}
 
-      return Reflect.set(target, property, slideIndex);
-    },
-  };
+			return Reflect.set(target, property, slideIndex);
+		},
+	};
 }
 
 indicatorLines.forEach((line, index) =>
-  line.addEventListener("click", (e) => {
-    e.stopPropagation();
-    watchedSlide.activeSlide = index;
-  })
+	line.addEventListener("click", (e) => {
+		e.stopPropagation();
+		watchedSlide.activeSlide = index;
+	})
 );
 
 function delay() {
-  timer = setTimeout(() => {
-    clearTimeout(timer);
-    timer = null;
-  }, 500);
+	timer = setTimeout(() => {
+		clearTimeout(timer);
+		timer = null;
+	}, 500);
 }
 
 function onWheel(e) {
-  clearTimeout(timeoutId);
+	clearTimeout(timeoutId);
 
-  const scrollAmountY = e.deltaY;
+	const scrollAmountY = e.deltaY;
 
-  const isScrollDown = scrollAmountY > 50;
-  const isScrollUp = scrollAmountY < -50;
+	const isScrollDown = scrollAmountY > 50;
+	const isScrollUp = scrollAmountY < -50;
 
-  if (isStart === false && timer === null) {
-    if (isScrollDown) {
-      isStart = true;
-      watchedSlide.activeSlide++;
+	if (isStart === false && timer === null) {
+		if (isScrollDown) {
+			isStart = true;
+			watchedSlide.activeSlide++;
 
-      delay();
-    }
+			delay();
+		}
 
-    if (isScrollUp) {
-      isStart = true;
-      watchedSlide.activeSlide--;
+		if (isScrollUp) {
+			isStart = true;
+			watchedSlide.activeSlide--;
 
-      delay();
-    }
-  }
+			delay();
+		}
+	}
 
-  timeoutId = setTimeout(function () {
-    isStart = false;
-  }, 25);
+	timeoutId = setTimeout(function () {
+		isStart = false;
+	}, 25);
 }
 
 function onDragStart(e) {
-  clientY1 = e.touches ? e.touches[0].clientY : e.clientY;
+	clientY1 = e.touches ? e.touches[0].clientY : e.clientY;
 }
 
 function onDragEnd(e) {
-  const clientY2 = e.changedTouches ? e.changedTouches[0].clientY : e.clientY;
-  const diff = clientY1 - clientY2;
+	const clientY2 = e.changedTouches ? e.changedTouches[0].clientY : e.clientY;
+	const diff = clientY1 - clientY2;
 
-  if (diff < -100) {
-    watchedSlide.activeSlide--;
-  }
-  if (diff > 100) {
-    watchedSlide.activeSlide++;
-  }
+	if (diff < -100) {
+		watchedSlide.activeSlide--;
+	}
+	if (diff > 100) {
+		watchedSlide.activeSlide++;
+	}
 }
 
 main.addEventListener("wheel", onWheel);
@@ -121,17 +121,53 @@ main.addEventListener("touchend", onDragEnd);
 main.addEventListener("mousedown", onDragStart);
 main.addEventListener("mouseup", onDragEnd);
 
-export { watchedSlide };
-
 // BurgerMenu //
 
+const hamburgerBodys = document.querySelector("#menu__body");
 const hamburgerBtn = document.querySelector("#wrapper__icon");
 const hamburgerBody = document.querySelector("#menu__body");
-import { closeBurger } from "./hamburger";
+const body = document.querySelector("body");
 
-hamburgerBody.querySelectorAll(".nav__link").forEach((link) => {
-  link.addEventListener("click", () => {
-    closeBurger();
-    watchedSlide.activeSlide = +link.dataset.sectionid;
-  });
+let isOpenBurger = false;
+let headerWasBlack = false;
+
+hamburgerBtn.addEventListener("click", () => {
+	isOpenBurger === true ? closeBurger() : openBurger();
+	console.log(headerWasBlack);
+});
+
+function openBurger() {
+	hamburgerBtn.classList.add("active");
+	hamburgerBody.classList.add("active");
+	body.classList.add("lock");
+	isOpenBurger = true;
+
+	if (header.classList.contains("black")) {
+		header.classList.remove("black");
+		headerWasBlack = true;
+	} else {
+		headerWasBlack = false;
+	}
+}
+
+function closeBurger() {
+	hamburgerBtn.classList.remove("active");
+	hamburgerBody.classList.remove("active");
+	body.classList.remove("lock");
+	isOpenBurger = false;
+
+	if (headerWasBlack) {
+		header.classList.add("black");
+	} else {
+		headerWasBlack = false;
+	}
+}
+
+hamburgerBodys.querySelectorAll(".nav__link").forEach((link) => {
+	link.addEventListener("click", () => {
+		closeBurger();
+		headerWasBlack = false;
+		watchedSlide.activeSlide = +link.dataset.sectionid;
+		console.log(headerWasBlack);
+	});
 });
